refactor: reorganize URL configuration commands into settings module
