+++ package.json
@@ -55,8 +55,6 @@
     "pnpm": ">=10"
   },
   "dependencies": {
-    "@tauri-apps/plugin-process": "~2.3.1",
-    "@tauri-apps/plugin-updater": "~2.10.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "next": "^16.2.6",
@@ -71,7 +69,7 @@
     "@tailwindcss/postcss": "^4.3.0",
     "@tauri-apps/api": "^2.11.0",
     "@tauri-apps/cli": "^2.11.1",
-    "@types/node": "^25.7.0",
+    "@types/node": "^25.8.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "shadcn": "^4.7.0",
